requirements.txt
django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==23.5
google-generativeai==0.3.2
google-genai==0.2.2
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-ratelimit==4.1.0